+++ src/main/java/tutsviews/lms/service/impl/ImageServiceImpl.java
@@ -2,6 +2,7 @@ package tutsviews.lms.service.impl;
 
 import java.util.List;
 
+import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,10 +17,13 @@ public class ImageServiceImpl implements ImageService {
 	@Autowired
 	ImageRepository imageRepository;
 	
+	@Autowired
+	Logger logger;
+	
 	
 	@Override
-	public void saveImage(Image image) {
-		imageRepository.save(image);
+	public void createImage(Image image) {
+		imageRepository.save(image); 
 	}
 
 	@Override
@@ -33,8 +37,15 @@ public class ImageServiceImpl implements ImageService {
 	}
 
 	@Override
-	public void deleteImage(int id) {
-		imageRepository.delete((long) id);
+	public boolean deleteImage(int id) {
+		if (!(imageRepository.findOne((long) id)==null)) {
+			imageRepository.delete((long) id);
+			logger.info("image with id "+id +" have been deleted.");
+			return true;
+		}else {
+			logger.error("cannot delete image with id "+id);
+			return false;
+		}
 	}
